requirements.txt
asgiref==3.6.0
attrs==22.2.0
Authlib==1.2.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.12.0
binance-connector==2.0.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.1
Django==4.1.7
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==22.3
django-filter==23.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
numpy==1.24.2
openpyxl==3.1.2
pandas==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-dateutil==2.8.2
pytz==2023.2
requests==2.28.2
service-identity==21.1.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
zope.interface==6.0
